# api/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/wg/configs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "List my WireGuard configs",
                "responses": {
                    "200": {
                        "description": "Configs listed successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.WGPeerListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/wg/configs/{id}/download": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "Download WireGuard config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "WireGuard config file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/wg/configs/{id}/revoke": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "Revoke WireGuard config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Revoked successfully",
                        "schema": {
                            "$ref": "#/definitions/core.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/wg/configs/{id}/rotate": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "Rotate WireGuard config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rotated successfully",
                        "schema": {
                            "$ref": "#/definitions/core.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "integer"
                },
                "details": {
                    "description": "Details contains detailed validation error information.\nThis field is only present when validation errors occur.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                },
                "reference": {
                    "description": "Reference returns the reference document which maybe useful to solve this error.",
                    "type": "string"
                }
            }
        },
        "core.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business success code.",
                    "type": "integer"
                },
                "message": {
                    "description": "Message contains the success message.",
                    "type": "string"
                }
            }
        },
...
        "v1.WGPeerListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.WGPeerResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "v1.WGPeerResponse": {
            "type": "object",
            "properties": {
                "allowed_ips": {
                    "type": "string"
                },
                "client_ip": {
                    "type": "string"
                },
                "client_public_key": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "persistent_keepalive": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}